+++ includes/CF7MessageFilter.php
@@ -24,7 +24,7 @@ class CF7MessageFilter
         // our constructor
         $this->blocked = get_option("kmcfmf_messages_blocked_today");
         //  $this->error_notice("hi there");
-        self::$version = '1.3.0';
+        self::$version = '1.3.1';
     }
 
     /**