+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Change Log
 ## v1.2.5 Changes
 1. Move log file to uploads folder
+2. Add auto clear message function
 
 ## v1.2.4 Changes
-1. Add auto clear message function
 1. Fix bug with some messages with links bypassing the checks
 
 ## v1.2.3 Changes